+++ core/src/main/java/com/graphhopper/routing/util/parsers/RoadsAverageSpeedParser.java
@@ -2,18 +2,17 @@ package com.graphhopper.routing.util.parsers;
 
 import com.graphhopper.reader.ReaderWay;
 import com.graphhopper.routing.ev.DecimalEncodedValue;
-import com.graphhopper.routing.ev.EncodedValueLookup;
 import com.graphhopper.routing.ev.EdgeIntAccess;
+import com.graphhopper.routing.ev.EncodedValueLookup;
 import com.graphhopper.routing.ev.VehicleSpeed;
 import com.graphhopper.storage.IntsRef;
-import com.graphhopper.util.PMap;
 
 public class RoadsAverageSpeedParser implements TagParser {
     private final DecimalEncodedValue avgSpeedEnc;
     private final double maxPossibleSpeed;
 
-    public RoadsAverageSpeedParser(EncodedValueLookup lookup, PMap properties) {
-        this(lookup.getDecimalEncodedValue(VehicleSpeed.key(properties.getString("name", "roads"))));
+    public RoadsAverageSpeedParser(EncodedValueLookup lookup) {
+        this(lookup.getDecimalEncodedValue(VehicleSpeed.key("roads")));
     }
 
     public RoadsAverageSpeedParser(DecimalEncodedValue avgSpeedEnc) {