+++ core/src/test/java/com/graphhopper/reader/osm/OSMReaderTest.java
@@ -31,7 +31,10 @@ import com.graphhopper.routing.OSMReaderConfig;
 import com.graphhopper.routing.ev.*;
 import com.graphhopper.routing.util.*;
 import com.graphhopper.routing.util.countryrules.CountryRuleFactory;
-import com.graphhopper.routing.util.parsers.*;
+import com.graphhopper.routing.util.parsers.CountryParser;
+import com.graphhopper.routing.util.parsers.OSMBikeNetworkTagParser;
+import com.graphhopper.routing.util.parsers.OSMMtbNetworkTagParser;
+import com.graphhopper.routing.util.parsers.OSMRoadAccessParser;
 import com.graphhopper.storage.*;
 import com.graphhopper.storage.index.LocationIndex;
 import com.graphhopper.storage.index.Snap;
@@ -142,14 +145,6 @@ public class OSMReaderTest {
     }
 
     @Test
-    public void testSort() {
-        GraphHopper hopper = new GraphHopperFacade(file1).setSortGraph(true).importOrLoad();
-        NodeAccess na = hopper.getBaseGraph().getNodeAccess();
-        assertEquals(10, na.getLon(findID(hopper.getLocationIndex(), 49, 10)), 1e-3);
-        assertEquals(51.249, na.getLat(findID(hopper.getLocationIndex(), 51.2492152, 9.4317166)), 1e-3);
-    }
-
-    @Test
     public void testOneWay() {
         GraphHopper hopper = new GraphHopperFacade(file2)
                 .setMinNetworkSize(0)
@@ -202,11 +197,7 @@ public class OSMReaderTest {
 
     @Test
     public void testFerry() {
-        GraphHopper hopper = new GraphHopperFacade(file2) {
-            @Override
-            public void cleanUp() {
-            }
-        }.importOrLoad();
+        GraphHopper hopper = new GraphHopperFacade(file2).importOrLoad();
         Graph graph = hopper.getBaseGraph();
 
         int n40 = AbstractGraphStorageTester.getIdOf(graph, 54.0);
@@ -228,11 +219,7 @@ public class OSMReaderTest {
 
     @Test
     public void testMaxSpeed() {
-        GraphHopper hopper = new GraphHopperFacade(file2) {
-            @Override
-            public void cleanUp() {
-            }
-        }.importOrLoad();
+        GraphHopper hopper = new GraphHopperFacade(file2).importOrLoad();
         Graph graph = hopper.getBaseGraph();
 
         int n60 = AbstractGraphStorageTester.getIdOf(graph, 56.0);
@@ -711,34 +698,18 @@ public class OSMReaderTest {
     @Test
     public void testTurnFlagCombination() {
         GraphHopper hopper = new GraphHopper();
-        hopper.setVehicleEncodedValuesFactory((name, config) -> {
-            if (name.equals("truck")) {
-                return VehicleEncodedValues.car(new PMap(config).putObject("name", "truck"));
-            } else {
-                return new DefaultVehicleEncodedValuesFactory().createVehicleEncodedValues(name, config);
-            }
-        });
-        hopper.setVehicleTagParserFactory((lookup, name, config) -> {
-            if (name.equals("truck")) {
-                return new VehicleTagParsers(
-                        new CarAccessParser(lookup.getBooleanEncodedValue(VehicleAccess.key("truck")), lookup.getBooleanEncodedValue(Roundabout.KEY), config, TransportationMode.HGV),
-                        new CarAverageSpeedParser(lookup.getDecimalEncodedValue(VehicleSpeed.key("truck")), lookup.getDecimalEncodedValue(FerrySpeed.KEY)),
-                        null
-                );
-            }
-            return new DefaultVehicleTagParserFactory().createParsers(lookup, name, config);
-        });
         hopper.setOSMFile(getClass().getResource("test-multi-profile-turn-restrictions.xml").getFile()).
                 setGraphHopperLocation(dir).
+                setVehiclesString("roads|transportation_mode=HGV|turn_costs=true").
                 setProfiles(
                         new Profile("bike").setVehicle("bike").setTurnCosts(true),
                         new Profile("car").setVehicle("car").setTurnCosts(true),
-                        new Profile("truck").setVehicle("truck").setTurnCosts(true)
+                        new Profile("truck").setVehicle("roads").setTurnCosts(true)
                 ).
                 importOrLoad();
         EncodingManager manager = hopper.getEncodingManager();
         BooleanEncodedValue carTCEnc = manager.getTurnBooleanEncodedValue(TurnRestriction.key("car"));
-        BooleanEncodedValue truckTCEnc = manager.getTurnBooleanEncodedValue(TurnRestriction.key("truck"));
+        BooleanEncodedValue truckTCEnc = manager.getTurnBooleanEncodedValue(TurnRestriction.key("roads"));
         BooleanEncodedValue bikeTCEnc = manager.getTurnBooleanEncodedValue(TurnRestriction.key("bike"));
 
         Graph graph = hopper.getBaseGraph();